cli,server: move runInitialSQL to the server package
This also enables us to get rid of the ill-advised `RunLocalSQL`
method on `Server`.

Release note: None
