+++ pkg/server/initial_sql.go
@@ -8,29 +8,27 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package cli
+package server
 
 import (
 	"context"
 	"fmt"
 
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver"
-	"github.com/cockroachdb/cockroach/pkg/server"
-	"github.com/cockroachdb/cockroach/pkg/sql"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/errors"
 )
 
-// runInitialSQL concerns itself with running "initial SQL" code when
+// RunInitialSQL concerns itself with running "initial SQL" code when
 // a cluster is started for the first time.
 //
 // The "startSingleNode" argument is true for `start-single-node`,
 // and `cockroach demo` with 2 nodes or fewer.
 // If adminUser is non-empty, an admin user with that name is
 // created upon initialization. Its password is then also returned.
-func runInitialSQL(
-	ctx context.Context, s *server.Server, startSingleNode bool, adminUser, adminPassword string,
+func (s *Server) RunInitialSQL(
+	ctx context.Context, startSingleNode bool, adminUser, adminPassword string,
 ) error {
 	newCluster := s.InitialStart() && s.NodeID() == kvserver.FirstNodeID
 	if !newCluster {
@@ -42,7 +40,7 @@ func runInitialSQL(
 		// For start-single-node, set the default replication factor to
 		// 1 so as to avoid warning messages and unnecessary rebalance
 		// churn.
-		if err := cliDisableReplication(ctx, s); err != nil {
+		if err := s.disableReplication(ctx); err != nil {
 			log.Ops.Errorf(ctx, "could not disable replication: %v", err)
 			return err
 		}
@@ -51,7 +49,7 @@ func runInitialSQL(
 	}
 
 	if adminUser != "" && !s.Insecure() {
-		if err := createAdminUser(ctx, s, adminUser, adminPassword); err != nil {
+		if err := s.createAdminUser(ctx, adminUser, adminPassword); err != nil {
 			return err
 		}
 	}
@@ -59,51 +57,53 @@ func runInitialSQL(
 	return nil
 }
 
-// createAdminUser creates an admin user with the given name.
-func createAdminUser(ctx context.Context, s *server.Server, adminUser, adminPassword string) error {
-	return s.RunLocalSQL(ctx,
-		func(ctx context.Context, ie *sql.InternalExecutor) error {
-			_, err := ie.Exec(
-				ctx, "admin-user", nil,
-				fmt.Sprintf("CREATE USER %s WITH PASSWORD $1", adminUser),
-				adminPassword,
-			)
-			if err != nil {
-				return err
-			}
-			// TODO(knz): Demote the admin user to an operator privilege with fewer options.
-			_, err = ie.Exec(ctx, "admin-user", nil, fmt.Sprintf("GRANT admin TO %s", tree.Name(adminUser)))
-			return err
-		})
+// RunInitialSQL implements cli.serverStartupInterface.
+func (s *SQLServerWrapper) RunInitialSQL(context.Context, bool, string, string) error {
+	return nil
 }
 
-// cliDisableReplication changes the replication factor on
+// createAdminUser creates an admin user with the given name.
+func (s *Server) createAdminUser(ctx context.Context, adminUser, adminPassword string) error {
+	ie := s.sqlServer.internalExecutor
+	_, err := ie.Exec(
+		ctx, "admin-user", nil,
+		fmt.Sprintf("CREATE USER %s WITH PASSWORD $1", adminUser),
+		adminPassword,
+	)
+	if err != nil {
+		return err
+	}
+	// TODO(knz): Demote the admin user to an operator privilege with fewer options.
+	_, err = ie.Exec(ctx, "admin-user", nil, fmt.Sprintf("GRANT admin TO %s", tree.Name(adminUser)))
+	return err
+}
+
+// disableReplication changes the replication factor on
 // all defined zones to become 1. This is used by start-single-node
 // and demo to define single-node clusters, so as to avoid
 // churn in the log files.
 //
 // The change is effected using the internal SQL interface of the
 // given server object.
-func cliDisableReplication(ctx context.Context, s *server.Server) error {
-	return s.RunLocalSQL(ctx,
-		func(ctx context.Context, ie *sql.InternalExecutor) (retErr error) {
-			it, err := ie.QueryIterator(ctx, "get-zones", nil,
-				"SELECT target FROM crdb_internal.zones")
-			if err != nil {
-				return err
-			}
-			// We have to make sure to close the iterator since we might return
-			// from the for loop early (before Next() returns false).
-			defer func() { retErr = errors.CombineErrors(retErr, it.Close()) }()
+func (s *Server) disableReplication(ctx context.Context) (retErr error) {
+	ie := s.sqlServer.internalExecutor
 
-			var ok bool
-			for ok, err = it.Next(ctx); ok; ok, err = it.Next(ctx) {
-				zone := string(*it.Cur()[0].(*tree.DString))
-				if _, err := ie.Exec(ctx, "set-zone", nil,
-					fmt.Sprintf("ALTER %s CONFIGURE ZONE USING num_replicas = 1", zone)); err != nil {
-					return err
-				}
-			}
+	it, err := ie.QueryIterator(ctx, "get-zones", nil,
+		"SELECT target FROM crdb_internal.zones")
+	if err != nil {
+		return err
+	}
+	// We have to make sure to close the iterator since we might return
+	// from the for loop early (before Next() returns false).
+	defer func() { retErr = errors.CombineErrors(retErr, it.Close()) }()
+
+	var ok bool
+	for ok, err = it.Next(ctx); ok; ok, err = it.Next(ctx) {
+		zone := string(*it.Cur()[0].(*tree.DString))
+		if _, err := ie.Exec(ctx, "set-zone", nil,
+			fmt.Sprintf("ALTER %s CONFIGURE ZONE USING num_replicas = 1", zone)); err != nil {
 			return err
-		})
+		}
+	}
+	return err
 }
