+++ pkg/server/server.go
@@ -1814,22 +1814,6 @@ func init() {
 	tracing.RegisterTagRemapping("n", "node")
 }
 
-// RunLocalSQL calls fn on a SQL internal executor on this server.
-// This is meant for use for SQL initialization during bootstrapping.
-//
-// The internal SQL interface should be used instead of a regular SQL
-// network connection for SQL initializations when setting up a new
-// server, because it is possible for the server to listen on a
-// network interface that is not reachable from loopback. It is also
-// possible for the TLS certificates to be invalid when used locally
-// (e.g. if the hostname in the cert is an advertised address that's
-// only reachable externally).
-func (s *Server) RunLocalSQL(
-	ctx context.Context, fn func(ctx context.Context, sqlExec *sql.InternalExecutor) error,
-) error {
-	return fn(ctx, s.sqlServer.internalExecutor)
-}
-
 // Insecure returns true iff the server has security disabled.
 func (s *Server) Insecure() bool {
 	return s.cfg.Insecure
